do_hhvm: only run quick tests for Debug build
All tests for Debug build takes >> 40 hours.
(11467.47+118267.02+14730.64+... = 144465.13s+++)

In contrast, Quick test for Debug build takes ~7 hours.
(601.35+7254.34+860.46+7555.38+938.75+7575.60 = 24785.88s)

This is on VM (4 vcpus) on linaro.cloud.

For reference, All tests for Release build takes ~30 hours.
(4241.27+30925.28+4345.27+31152.91+4410.15+31176.07 = 106250.95s)

With build time of ~2 hours for each build variant, the total run-time
of this script should be within the bounds of 2 days.